Sanitize analytics and configure Supabase
